Move integration test into ide subproject
